Monviso.prox — Function

prox(y::AbstractVector{VariableRef}, model::Model; g=nothing, norm_cone=MOI.SecondOrderCone)

The proximal operator closure.

Arguments

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

g::Function=nothing - a function of the form x::AbstractVector{VariableRef} -> Real, i.e., the scalar function $g : \mathbb{R}^n \to \mathbb{R}$.

norm_cone=MOI.SecondOrderCone - the cone related to the norm characterizing the proximal operator.

Monviso.prox_gradient — Function

prox_gradient(F, y::AbstractVector{VariableRef}, model::Model; g=nothing, analytical_prox=nothing, norm_cone=MOI.SecondOrderCone)

The proximal gradient closure.

Description

Given a constant step-size $\chi > 0$ and an initial vector $\mathbf{x}_0 \in \mathbb{R}^n$, the basic $k$-th iterate of the proximal gradient (PG) algorithm is [1]:

\[\mathbf{x}_{k+1} = \text{prox}_{g,\mathcal{S}}(\mathbf{x}_k - \chi \mathbf{F}(\mathbf{x}_k))\]

where $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of PG is guaranteed for Lipschitz strongly monotone operators, with monotone constant $\mu > 0$ and Lipshitz constants $L < +\infty$, when $\chi \in (0, 2\mu/L^2)$.

Arguments

F - a function of the form (x::AbstractVector, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

g::Function=nothing - a function of the form x::AbstractVector{VariableRef} -> Real, i.e., the scalar function $g : \mathbb{R}^n \to \mathbb{R}$.

analytical_prox::Function=nothing - the analytical form of the proximal operator for the given set. If provided, it replaces of prox.

norm_cone=MOI.SecondOrderCone - the cone related to the norm characterizing the proximal operator.

Monviso.forward_backward_forward — Function

forward_backward_forward(F, y::AbstractVector{VariableRef}, model::Model; g=nothing, analytical_prox=nothing, norm_cone=MOI.SecondOrderCone)

The forward-backward-forward closure.

Description

Given a constant step-size $\chi > 0$ and an initial vector $\mathbf{x}_0 \in \mathbb{R}^n$, the $k$-th iterate of Forward-Backward-Forward (FBF) algorithm is [1]:

\[\begin{align*} \mathbf{y}_k &= \text{prox}_{\mathcal{S}}(\mathbf{x}_k - \chi F(\mathbf{x}_k)) \\ \mathbf{x}_{k+1} &= \mathbf{y}_k - \chi (F(\mathbf{y}_k) - \chi F(\mathbf{x}_k)) \end{align*}\]

where $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of the FBF algorithm is guaranteed for Lipschitz monotone operators, with Lipschitz constant $L < +\infty$, when $\chi \in \left(0,\frac{1}{L}\right)$.

Arguments

F - a function of the form (x::AbstractVector, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

g::Function=nothing - a function of the form x::AbstractVector{VariableRef} -> Real, i.e., the scalar function $g : \mathbb{R}^n \to \mathbb{R}$.

analytical_prox::Function=nothing - the analytical form of the proximal operator for the given set. If provided, it replaces of prox.

norm_cone=MOI.SecondOrderCone - the cone related to the norm characterizing the proximal operator.

Monviso.extragradient — Function

extragradient(F, y::AbstractVector{VariableRef}, model::Model; g=nothing, analytical_prox=nothing, norm_cone=MOI.SecondOrderCone)

The extragradient closure

Description

Given a constant step-size $\chi > 0$ and an initial vector $\mathbf{x}_0 \in \mathbb{R}^n$, the $k$-th iterate of the extragradient algorithm (EG) is[1]:

\[\begin{align*} \mathbf{y}_k &= \text{prox}_{\mathcal{S}}(\mathbf{x}_k - \chi \mathbf{F}(\mathbf{x}_k)) \\ \mathbf{x}_{k+1} &= \text{prox}_{\mathcal{S}}(\mathbf{y}_k - \chi \mathbf{F}(\mathbf{x}_k)) \end{align*}\]

where $g : \mathbb{R}^n \to \mathbb{R}$ is a scalar convex (possibly non-smooth) function, while $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of the EGD algorithm is guaranteed for Lipschitz monotone operators, with Lipschitz constant $L < +\infty$, when $\chi \in \left(0,\frac{1}{L}\right)$.

Arguments

F - a function of the form (x::AbstractVector, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

g::Function=nothing - a function of the form x::AbstractVector{VariableRef} -> Real, i.e., the scalar function $g : \mathbb{R}^n \to \mathbb{R}$.

analytical_prox::Function=nothing - the analytical form of the proximal operator for the given set. If provided, it replaces of prox.

norm_cone=MOI.SecondOrderCone - the cone related to the norm characterizing the proximal operator.

Monviso.popov — Function

popov(F, y::AbstractVector{VariableRef}, model::Model; g=nothing, analytical_prox=nothing, norm_cone=MOI.SecondOrderCone)

The Popov's method closure

Given a constant step-size $\chi > 0$ and an initial vectors $\mathbf{x}_0,\mathbf{y}_0 \in \mathbb{R}^n$, the $k$-th iterate of Popov's Method (PM) is[1]:

\[\begin{align*} \mathbf{y}_{k+1} &= \text{prox}_{g,\mathcal{S}}(\mathbf{x}_k - \chi \mathbf{F}(\mathbf{y}_k)) \\ \mathbf{x}_{k+1} &= \text{prox}_{g,\mathcal{S}}(\mathbf{y}_{k+1} - \chi \mathbf{F}(\mathbf{x}_k)) \end{align*}\]

where $g : \mathbb{R}^n \to \mathbb{R}$ is a scalar convex (possibly non-smooth) function, while $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of PM is guaranteed for Lipschitz monotone operators, with Lipschitz constant $L < +\infty$, when $\chi \in \left(0,\frac{1}{2L}\right)$.

Arguments

F - a function of the form (x::AbstractVector, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

g::Function=nothing - a function of the form x::AbstractVector{VariableRef} -> Real, i.e., the scalar function $g : \mathbb{R}^n \to \mathbb{R}$.

analytical_prox::Function=nothing - the analytical form of the proximal operator for the given set. If provided, it replaces of prox.

norm_cone=MOI.SecondOrderCone - the cone related to the norm characterizing the proximal operator.

1Nemirovskij, A. S., & Yudin, D. B. (1983). Problem complexity and method efficiency in optimization.

2Tseng, P. (2000). A modified forward-backward splitting method for maximal monotone mappings. SIAM Journal on Control and Optimization, 38(2), 431-446.

3Korpelevich, G. M. (1976). The extragradient method for finding saddle points and other problems. Matecon, 12, 747-756.

4Popov, L.D. A modification of the Arrow-Hurwicz method for search of saddle points. Mathematical Notes of the Academy of Sciences of the USSR 28, 845–848 (1980)