Monviso.proj — Function

proj(optimizer, x_func::Function, set::SetType=default_set; norm_cone::DataType=MOI.SecondOrderCone, silent::Bool=true)

The projection operator closure.

Arguments

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

norm_cone::DataType=MOI.SecondOrderCone - the cone related to the norm characterizing the projection.

Monviso.proj_gradient — Function

proj_gradient(optimizer, F::Function, x_func::Function, set::SetType=default_set; analytical_proj::Union{Nothing, Function}=nothing, norm_cone::DataType=MOI.SecondOrderCone, silent::Bool=true)

The projected gradient closure.

Description

Given a constant step-size $\chi > 0$ and an initial vector $\mathbf{x}_0 \in \mathbb{R}^n$, the basic $k$-th iterate of the projected gradient (PG) algorithm is [1]:

\[\mathbf{x}_{k+1} = \text{proj}_{\mathcal{S}}(\mathbf{x}_k - \chi \mathbf{F}(\mathbf{x}_k))\]

where $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of PG is guaranteed for Lipschitz strongly monotone operators, with monotone constant $\mu > 0$ and Lipshitz constants $L < +\infty$, when $\chi \in (0, 2\mu/L^2)$.

Arguments

F - a function of the form (x, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

analytical_proj::=nothing - the analytical form of the projection of the given set. If provided, it replaces of proj.

norm_cone::DataType=MOI.SecondOrderCone - the cone related to the norm characterizing the projection.

References

[1] Nemirovskij, A. S., & Yudin, D. B. (1983). Problem complexity and method efficiency in optimization.

Monviso.forward_backward_forward — Function

forward_backward_forward(optimizer, F::Function, x_func::Function, set::SetType=default_set; analytical_proj::Union{Nothing, Function}=nothing, norm_cone::DataType=MOI.SecondOrderCone, silent::Bool=true)

The forward-backward-forward closure.

Description

Given a constant step-size $\chi > 0$ and an initial vector $\mathbf{x}_0 \in \mathbb{R}^n$, the $k$-th iterate of Forward-Backward-Forward (FBF) algorithm is [1]:

\[\begin{align*} \mathbf{y}_k &= \text{proj}_{\mathcal{S}}(\mathbf{x}_k - \chi F(\mathbf{x}_k)) \\ \mathbf{x}_{k+1} &= \mathbf{y}_k - \chi (F(\mathbf{y}_k) - \chi F(\mathbf{x}_k)) \end{align*}\]

where $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$ is the VI mapping. The convergence of the FBF algorithm is guaranteed for Lipschitz monotone operators, with Lipschitz constant $L < +\infty$, when $\chi \in \left(0,\frac{1}{L}\right)$.

Arguments

F - a function of the form (x, params...) -> AbstractVector of the same lenght of x, i.e., the VI mapping $\mathbf{F} : \mathbb{R}^n \to \mathbb{R}^n$. Term params collects optional arguments that characterize F and might change at each iteration.

y::AbstractVector{VariableRef} - the container of JuMP.VariableRef associated to model, i.e., $\mathbf{y}$.

model::Model - the JuMP.Model describing the projection set $\mathcal{S}$.

Keywords

analytical_proj::=nothing - the analytical form of the projection of the given set. If provided, it replaces of proj.

norm_cone::DataType=MOI.SecondOrderCone - the cone related to the norm characterizing the projection.

References

[1] Tseng, P. (2000). A modified forward-backward splitting method for maximal monotone mappings. SIAM Journal on Control and Optimization, 38(2), 431-446.